+++ source/js/search-parameters/common-descriptions.js
@@ -1,7 +1,9 @@
 import { FhirBatchQuery } from '../common/fhir-batch-query';
 import {
   encodeFhirSearchParameter,
-  getAutocompleterById
+  getAutocompleterById,
+  capitalize,
+  toggleCssClass
 } from '../common/utils';
 import definitionsIndex from './definitions/index.json';
 
@@ -165,16 +167,41 @@ function stringParameterDescription({ placeholder, column, name }) {
  * @param {string} description - title for input field,
  * @param {Object} column - HTML table column name
  * @param {string} name - name of search parameter to construct result query string
+ * @param {string} elementPath - resource element path
+ * @param {string} resourceType - resource type, e.g. 'Patient', 'Encounter'
  * @return {Object}
  */
-function dateParameterDescription({ name, column, description }) {
+function dateParameterDescription({
+  name,
+  column,
+  description,
+  elementPath,
+  resourceType
+}) {
   return {
     column,
     getControlsHtml: (searchItemId) => {
       const title = (description && description.replace(/"/g, '&quot;')) || '';
       return `\
-<span>from</span><input type="date" id="${searchItemId}-${name}-from" placeholder="no limit" title="${title}">
-<span>to</span><input type="date" id="${searchItemId}-${name}-to" placeholder="no limit" title="${title}">`;
+<span>from</span><input type="date" id="${searchItemId}-${name}-from" placeholder="yyyy-mm-dd" title="${title}">
+<span>to</span><input type="date" id="${searchItemId}-${name}-to" placeholder="yyyy-mm-dd" title="${title}">`;
+    },
+    attachControls: (searchItemId) => {
+      const fromId = `${searchItemId}-${name}-from`;
+      const toId = `${searchItemId}-${name}-to`;
+
+      if (elementPath && resourceType) {
+        switchLoadingStatus(searchItemId, true);
+        Promise.all(
+          [
+            loadDate(fromId, resourceType, name, elementPath, true),
+            loadDate(toId, resourceType, name, elementPath, false)
+          ].map((i) =>
+            // Convert reject to resolve to emulate Promise.allSettled behaviour (for IE11)
+            i.catch((error) => error)
+          )
+        ).then(() => switchLoadingStatus(searchItemId, false));
+      }
     },
     getCondition: (searchItemId) => {
       const from = document.getElementById(`${searchItemId}-${name}-from`)
@@ -187,6 +214,78 @@ function dateParameterDescription({ name, column, description }) {
       );
     }
   };
+}
+
+/**
+ * Load default minimum/maximum date value to input field from database.
+ * @param {string} inputId - input field id
+ * @param {string} resourceType - resource type
+ * @param {string} paramName - search parameter name
+ * @param {string} resourceElementPath - resource element path
+ * @param {boolean} min - true/false to fill input with minimum/maximum value
+ */
+function loadDate(inputId, resourceType, paramName, resourceElementPath, min) {
+  const elementPath = resourceElementPath.split('.');
+
+  return getCurrentClient()
+    .getWithCache(
+      `${resourceType}?_count=1&_elements=${elementPath[0]}&_sort=${
+        (min ? '' : '-') + paramName
+      }`
+    )
+    .then(({ status, data }) => {
+      if (status === 200 && data.entry && data.entry.length) {
+        let value = getValueByPath(data.entry[0].resource, elementPath);
+        if (value && (value.start || value.end)) {
+          value = min ? value.start || value.end : value.end || value.start;
+        }
+        const date = /^(\d{4})(-\d{2}-\d{2}|$)/.test(value)
+          ? `${RegExp.$1}${RegExp.$2 || '-01-01'}`
+          : null;
+        if (date) {
+          updateInputIfExist(inputId, date);
+        }
+      }
+    });
+}
+
+/**
+ * Returns value from Object by path
+ * @param {Object} value - input Object
+ * @param {Array} path - array of property names
+ * @return {*}
+ */
+function getValueByPath(value, path) {
+  let i = 0;
+  while (value && i < path.length) {
+    value = value && value[path[i]];
+    i++;
+  }
+  return value;
+}
+
+/**
+ * Changes input field value if exists
+ * @param {string} inputId
+ * @param {string} value
+ */
+function updateInputIfExist(inputId, value) {
+  const input = document.getElementById(inputId);
+  if (input) {
+    input.value = value;
+  }
+}
+
+/**
+ * Shows/hides search parameter controls with a spinner depending on the boolean parameter "loading".
+ * @param {string} searchItemId
+ * @param {boolean} loading
+ */
+function switchLoadingStatus(searchItemId, loading) {
+  const content = document.getElementById(searchItemId + '_content');
+  if (content) {
+    toggleCssClass(content, 'spinner', loading);
+  }
 }
 
 /**
@@ -248,9 +347,7 @@ export function defaultParameters(
 
   return definitions.resources[resourceType].reduce((_parameters, item) => {
     if (skip.indexOf(item.name) === -1) {
-      const displayName =
-        item.name.charAt(0).toUpperCase() +
-        item.name.substring(1).replace(/-/g, ' ');
+      const displayName = capitalize(item.name).replace(/-/g, ' ');
       const placeholder = item.description;
       const name = item.name;
 
@@ -260,7 +357,12 @@ export function defaultParameters(
           _parameters[displayName] = dateParameterDescription({
             description: item.description,
             name,
-            column: searchNameToColumn[name] || name
+            column: searchNameToColumn[name] || name,
+            elementPath: getElementPathFromExpression(
+              resourceType,
+              item.expression
+            ),
+            resourceType
           });
           break;
         case 'boolean':
@@ -293,6 +395,24 @@ export function defaultParameters(
 
     return _parameters;
   }, {});
+}
+
+/**
+ * Determines resource element path by expression from resource search parameter specification.
+ * @param {string} resourceType
+ * @param {string} expression
+ * @return {string}
+ */
+function getElementPathFromExpression(resourceType, expression) {
+  let s = new RegExp(`${resourceType}\\.([^\\s]*)`).test(expression)
+    ? RegExp.$1
+    : null;
+
+  if (/^(.*).as\((.*)\)$/.test(s)) {
+    return RegExp.$1 + capitalize(RegExp.$2);
+  }
+
+  return s;
 }
 
 /**
@@ -348,18 +468,25 @@ export function valueSetsParameters(descriptions, searchNameToColumn) {
  * @param {Array[]} descriptions - an array of descriptions, each of which is an array containing the following elements:
  *                  displayName - parameter display name,
  *                  name - name of search parameter to construct result query string
+ *                  elementPath - resource element path
  * @param {Object} searchNameToColumn - mapping from search parameter names to HTML table column names
+ * @param {string} resourceType - resource type, e.g. 'Patient', 'Encounter'
  * @return {Object}
  */
-export function dateParameters(descriptions, searchNameToColumn) {
-  return descriptions.reduce((_parameters, [displayName, name]) => {
-    _parameters[displayName] = dateParameterDescription({
-      name,
-      column: searchNameToColumn[name] || name
-    });
+export function dateParameters(descriptions, searchNameToColumn, resourceType) {
+  return descriptions.reduce(
+    (_parameters, [displayName, name, elementPath]) => {
+      _parameters[displayName] = dateParameterDescription({
+        name,
+        column: searchNameToColumn[name] || name,
+        elementPath,
+        resourceType
+      });
 
-    return _parameters;
-  }, {});
+      return _parameters;
+    },
+    {}
+  );
 }
 
 /**